Treasury Lot 0: widen daily_rates precision to NUMERIC(10,4)
The four daily_rates rate columns were INTEGER, which prevented
entering fractional rates. Rates are expressed as "CNY per 1M XAF"
and typically sit around 11 000, so integer-step granularity
translates to ~9 000 XAF on a 100M XAF deal.

Migration:
- ALTER the four rate_* columns to NUMERIC(10,4) (non-destructive
  cast for existing integer values).
- DROP + recreate create_daily_rates() with NUMERIC params and
  added positivity guard.
- Recreate calculate_final_rate() with v_base_rate as NUMERIC so
  fractional rates from the columns are not silently truncated.
- Bump rounding of final_rate to 4 decimals in the returned JSONB.

Frontend:
- RateSetTab.tsx: replace 11 parseInt() calls with parseFloat() so
  the input UI no longer strips decimals before submission.

Verified: type-check passes, build succeeds, paymentRateLogic test
suite passes (14/14).

+++ src/mobile/screens/rates/tabs/RateSetTab.tsx
@@ -33,10 +33,10 @@ export function RateSetTab({ currentRate }: RateSetTabProps) {
 
   // Helper: build the rates object passed to the Edge Function
   const flyerRates = () => ({
-    alipay: parseInt(rates.alipay)   || currentRate?.rate_alipay   || 0,
-    wechat: parseInt(rates.wechat)   || currentRate?.rate_wechat   || 0,
-    bank:   parseInt(rates.virement) || currentRate?.rate_virement || 0,
-    cash:   parseInt(rates.cash)     || currentRate?.rate_cash     || 0,
+    alipay: parseFloat(rates.alipay)   || currentRate?.rate_alipay   || 0,
+    wechat: parseFloat(rates.wechat)   || currentRate?.rate_wechat   || 0,
+    bank:   parseFloat(rates.virement) || currentRate?.rate_virement || 0,
+    cash:   parseFloat(rates.cash)     || currentRate?.rate_cash     || 0,
   });
 
   const getEffectiveAt = (): string => {
@@ -59,10 +59,10 @@ export function RateSetTab({ currentRate }: RateSetTabProps) {
 
   const handleApply = () => {
     createRates.mutate({
-      rate_cash: parseInt(rates.cash) || 0,
-      rate_alipay: parseInt(rates.alipay) || 0,
-      rate_wechat: parseInt(rates.wechat) || 0,
-      rate_virement: parseInt(rates.virement) || 0,
+      rate_cash: parseFloat(rates.cash) || 0,
+      rate_alipay: parseFloat(rates.alipay) || 0,
+      rate_wechat: parseFloat(rates.wechat) || 0,
+      rate_virement: parseFloat(rates.virement) || 0,
       effective_at: getEffectiveAt(),
     });
   };
@@ -306,10 +306,10 @@ export function RateSetTab({ currentRate }: RateSetTabProps) {
         <div className="px-4 pb-3 overflow-hidden flex justify-center">
           <div style={{ transform: 'scale(0.172)', transformOrigin: 'top center', height: Math.round(2560 * 0.172), pointerEvents: 'none' }}>
             <RateFlyer
-              alipay={parseInt(rates.alipay) || currentRate?.rate_alipay || 0}
-              wechat={parseInt(rates.wechat) || currentRate?.rate_wechat || 0}
-              bank={parseInt(rates.virement) || currentRate?.rate_virement || 0}
-              cash={parseInt(rates.cash) || currentRate?.rate_cash || 0}
+              alipay={parseFloat(rates.alipay) || currentRate?.rate_alipay || 0}
+              wechat={parseFloat(rates.wechat) || currentRate?.rate_wechat || 0}
+              bank={parseFloat(rates.virement) || currentRate?.rate_virement || 0}
+              cash={parseFloat(rates.cash) || currentRate?.rate_cash || 0}
               theme={flyerDark ? 'dark' : 'light'}
             />
           </div>
